Mutualize checkCaptcha
