+++ server/graphql/v1/mutations/orders.js
@@ -20,6 +20,7 @@ import { getOrCreateGuestProfile } from '../../../lib/guest-accounts';
 import logger from '../../../lib/logger';
 import * as libPayments from '../../../lib/payments';
 import recaptcha from '../../../lib/recaptcha';
+import { checkCaptcha } from '../../../lib/check-captcha';
 import { getChargeRetryCount, getNextChargeAndPeriodStartDates } from '../../../lib/recurring-contributions';
 import { checkGuestContribution, checkOrdersLimit, cleanOrdersLimit } from '../../../lib/security/limit';
 import { orderFraudProtection } from '../../../lib/security/order';
@@ -101,45 +102,6 @@ const checkAndUpdateProfileInfo = async (order, fromAccount, isGuest, currency) 
     }
   }
 };
-
-async function checkCaptcha(order, remoteUser, reqIp) {
-  const requestedProvider = order.guestInfo?.captcha?.provider;
-  const isCaptchaEnabled = parseToBoolean(config.captcha?.enabled);
-  const isCreditCardOrder = order.paymentMethod?.type === PAYMENT_METHOD_TYPE.CREDITCARD;
-
-  // We're just enforcing Captcha if it's enabled and if the order is using Credit Card
-  if (!isCaptchaEnabled || !isCreditCardOrder) {
-    return;
-  }
-
-  if (!order.guestInfo?.captcha?.token) {
-    throw new BadRequest('You need to inform a valid captcha token');
-  }
-  let response;
-  if (requestedProvider === CAPTCHA_PROVIDERS.HCAPTCHA && config.hcaptcha?.secret) {
-    response = await hcaptcha.verify(
-      config.hcaptcha.secret,
-      order.guestInfo.captcha.token,
-      reqIp,
-      config.hcaptcha.sitekey,
-    );
-  } else if (
-    requestedProvider === CAPTCHA_PROVIDERS.RECAPTCHA &&
-    config.recaptcha &&
-    parseToBoolean(config.recaptcha.enable)
-  ) {
-    response = await recaptcha.verify(order.guestInfo.captcha.token, reqIp);
-  } else {
-    throw new BadRequest('Could not find requested Captcha provider', undefined, order.guestInfo?.captcha);
-  }
-
-  if (response.success !== true) {
-    logger.warn('Captcha verification failed:', response);
-    throw new BadRequest('Captcha verification failed');
-  }
-
-  return response;
-}
 
 /**
  * Check the taxes for order, returns the tax info
@@ -415,7 +377,16 @@ export async function createOrder(order, req) {
       } else {
         // Create or retrieve guest profile from GUEST_TOKEN
         const creationRequest = { ip: reqIp, userAgent, mask: reqMask };
-        captchaResponse = await checkCaptcha(order, remoteUser, reqIp);
+
+        // We're just enforcing Captcha if the order is using Credit Card
+        const isCreditCardOrder = order.paymentMethod?.type === PAYMENT_METHOD_TYPE.CREDITCARD;
+        if (isCreditCardOrder) {
+          try {
+            captchaResponse = await checkCaptcha(order.guestInfo?.captcha, reqIp);
+          } catch (err) {
+            throw new BadRequest(err.message, undefined, order.guestInfo?.captcha);
+          }
+        }
         const guestProfile = await getOrCreateGuestProfile(order.guestInfo, creationRequest);
         if (!canUseFeature(guestProfile.user, FEATURE.ORDER)) {
           throw new FeatureNotAllowedForUser();
